backend/requirements.txt
asgiref==3.2.10
requests==2.26.0
django==2.2.16
django-filter==2.4.0
djangorestframework==3.12.4
django-cors-headers==3.11.0
djoser==2.1.0
gunicorn==20.0.4
python-dotenv==0.21.1
psycopg2-binary==2.8.6
PyJWT==2.1.0
pytz==2020.1
sqlparse==0.3.1
pillow==9.4.0
reportlab==3.6.12
